+++ projects/ngx-material/src/directives/ui-wrapper.directive.ts
@@ -4,27 +4,33 @@ import {
 import {Subscription} from "rxjs";
 import {BaseUIScopeContext, UIScopeContext} from "../models/ui-scope";
 import {CdkScrollable, ScrollDispatcher} from "@angular/cdk/overlay";
+import {IScrollContext, ScrollContext} from "@juulsgaard/ngx-tools";
+import {takeUntilDestroyed} from "@angular/core/rxjs-interop";
 
 @Directive({
   selector: '[uiWrapper],ui-wrapper',
   standalone: true,
-  providers: [{provide: UIScopeContext, useExisting: forwardRef(() => UiWrapperDirective)}],
+  providers: [
+    {provide: UIScopeContext, useExisting: forwardRef(() => UiWrapperDirective)},
+    ScrollContext.Provide(() => UiWrapperDirective)
+  ],
   host: {'[class.ui-wrapper]': 'true'}
 })
-export class UiWrapperDirective extends BaseUIScopeContext implements OnInit, OnDestroy {
+export class UiWrapperDirective extends BaseUIScopeContext implements OnInit, OnDestroy, IScrollContext {
 
   private sub?: Subscription;
 
   @HostBinding('class')
   wrapperClass: string[] = [];
+  scrollable = true;
 
   private context: UIScopeContext;
-  private scrollable: CdkScrollable;
+  cdkScrollable: CdkScrollable;
 
   constructor(
     private changes: ChangeDetectorRef,
     private scrollDispatcher: ScrollDispatcher,
-    elementRef: ElementRef<HTMLElement>,
+    readonly element: ElementRef<HTMLElement>,
     ngZone: NgZone,
   ) {
     const context = inject(UIScopeContext, {skipSelf: true});
@@ -32,21 +38,24 @@ export class UiWrapperDirective extends BaseUIScopeContext implements OnInit, On
 
     this.context = context;
 
-    this.scrollable = new CdkScrollable(elementRef, scrollDispatcher, ngZone);
+    this.cdkScrollable = new CdkScrollable(element, scrollDispatcher, ngZone);
+
+    this.context.hasChildren$.pipe(takeUntilDestroyed()).subscribe(hasChildren => {
+      this.scrollable = !hasChildren;
+      if (!hasChildren) this.scrollDispatcher.register(this.cdkScrollable);
+      else this.scrollDispatcher.deregister(this.cdkScrollable);
+    });
   }
 
   ngOnInit() {
     this.sub = this.context.registerWrapper(x => {
       this.wrapperClass = x.classes;
       this.changes.detectChanges();
-      if (x.scrollable) this.scrollDispatcher.register(this.scrollable);
-      else this.scrollDispatcher.deregister(this.scrollable);
     });
   }
 
   ngOnDestroy() {
     this.sub?.unsubscribe();
-    this.scrollDispatcher.deregister(this.scrollable);
+    this.scrollDispatcher.deregister(this.cdkScrollable);
   }
-
 }